Currency, Price, TradingPair refactor
- Extracts `Currency` into the `delphi::currency` module
- Extracts `Price` and `PriceQuantity` into the `delphi::price` module
- Renames `Pair` => `TradingPair`; extracts into `delphi::trading_pair`
- `Price` improvements:
  - Makes inner `Decimal` non-pub so constructor can maintain invariant
  - Adds `Add`, `Mul`, and `Div` impls so we can do math on `Price`
  - Adds a `Deref` impl as a way to get to the inner `Decimal `easily
  - Adds `From` and `TryInto` impls for `Decimal`
  - Impls `Copy` on `Price` (inner `Decimal` is `Copy` so why not)
- Updates Binance source to return `Price`
