+++ src/sources.rs
@@ -6,21 +6,8 @@ pub mod coinone;
 pub mod gdac;
 pub mod gopax;
 
-use crate::{
-    error::{Error, ErrorKind},
-    prelude::*,
-};
-use percent_encoding::{utf8_percent_encode, NON_ALPHANUMERIC};
-use rust_decimal::{prelude::*, Decimal};
-use serde::{
-    de::{self, Error as _},
-    ser, Deserialize, Serialize,
-};
-use std::{
-    cmp::Ordering,
-    fmt::{self, Display},
-    str::FromStr,
-};
+use crate::{Error, Price, PriceQuantity};
+use rust_decimal::Decimal;
 
 /// User-Agent to send in HTTP request
 pub const USER_AGENT: &str = "iqlusion delphi";
@@ -43,7 +30,7 @@ pub fn weighted_avg_ask<T: AskBook>(asks: &T) -> Result<Price, Error> {
     let mut price_sum_product = Decimal::from(0u8);
     let mut total = Decimal::from(0u8);
     for ask in asks {
-        price_sum_product += ask.price.0 * ask.quantity;
+        price_sum_product += *ask.price * ask.quantity;
         total += ask.quantity;
     }
 
@@ -57,7 +44,7 @@ pub fn weighted_avg_bid<T: BidBook>(bids: &T) -> Result<Price, Error> {
     let mut price_sum_product = Decimal::from(0u8);
     let mut total = Decimal::from(0u8);
     for bid in bids {
-        price_sum_product += bid.price.0 * bid.quantity;
+        price_sum_product += *bid.price * bid.quantity;
         total += bid.quantity;
     }
 
@@ -69,238 +56,19 @@ pub fn weighted_avg_bid<T: BidBook>(bids: &T) -> Result<Price, Error> {
 pub fn lowest_ask<T: AskBook>(asks: &T) -> Result<Price, Error> {
     let mut asks = asks.asks()?;
     asks.sort();
-    Ok(asks.first().unwrap().price.clone())
+    Ok(asks.first().unwrap().price)
 }
 
 /// Lowest bid price
 pub fn highest_bid<T: BidBook>(bids: &T) -> Result<Price, Error> {
     let mut bids = bids.bids()?;
     bids.sort();
-    Ok(bids.last().unwrap().price.clone())
+    Ok(bids.last().unwrap().price)
 }
 
 /// Midpoint of highest ask and lowest bid price
 pub fn midpoint<T: AskBook + BidBook>(book: &T) -> Result<Price, Error> {
     let lowest_ask = lowest_ask(book)?;
     let highest_bid = highest_bid(book)?;
-    Price::new((lowest_ask.0 + highest_bid.0) / Decimal::from(2))
-}
-
-/// Quoted prices and quantities as sourced from the order book
-#[derive(Eq)]
-pub struct PriceQuantity {
-    ///Price
-    pub price: Price,
-
-    ///Quantity
-    pub quantity: Decimal,
-}
-
-impl Ord for PriceQuantity {
-    fn cmp(&self, other: &Self) -> Ordering {
-        self.price.cmp(&other.price)
-    }
-}
-
-impl PartialOrd for PriceQuantity {
-    fn partial_cmp(&self, other: &Self) -> Option<Ordering> {
-        Some(self.cmp(other))
-    }
-}
-
-impl PartialEq for PriceQuantity {
-    fn eq(&self, other: &Self) -> bool {
-        self.price == other.price
-    }
-}
-
-/// Currencies for use in trading pairs
-#[derive(Clone, Debug, Eq, PartialEq, PartialOrd, Ord)]
-pub enum Currency {
-    /// Cosmos Atom
-    Atom,
-
-    /// Binance Coin
-    Bnb,
-
-    /// Binance KRW (won) stablecoin
-    Bkrw,
-
-    /// Bitcoin
-    Btc,
-
-    /// Binance USD stablecoin
-    Busd,
-
-    /// Ethereum
-    Eth,
-
-    /// Euro
-    Eur,
-
-    /// UK Pounds
-    Gbp,
-
-    /// South Korean won
-    Krw,
-
-    /// Terra Luna
-    Luna,
-
-    /// US dollars
-    Usd,
-
-    /// Circle stablecoin
-    Usdc,
-
-    /// Tether USDT stablecoin
-    Usdt,
-
-    /// Other (open-ended)
-    Other(String),
-}
-
-impl Display for Currency {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        f.write_str(match self {
-            Currency::Atom => "ATOM",
-            Currency::Bnb => "BNB",
-            Currency::Bkrw => "BKRW",
-            Currency::Btc => "BTC",
-            Currency::Busd => "BUSD",
-            Currency::Eth => "ETH",
-            Currency::Eur => "EUR",
-            Currency::Gbp => "GBP",
-            Currency::Krw => "KRW",
-            Currency::Luna => "LUNA",
-            Currency::Usd => "USD",
-            Currency::Usdc => "USDC",
-            Currency::Usdt => "USDT",
-            Currency::Other(other) => other.as_ref(),
-        })
-    }
-}
-
-impl FromStr for Currency {
-    type Err = Error;
-
-    fn from_str(s: &str) -> Result<Self, Error> {
-        Ok(match s.to_ascii_uppercase().as_ref() {
-            "ATOM" => Currency::Atom,
-            "BNB" => Currency::Bnb,
-            "BKRW" => Currency::Bkrw,
-            "BTC" => Currency::Btc,
-            "BUSD" => Currency::Busd,
-            "ETH" => Currency::Eth,
-            "EUR" => Currency::Eur,
-            "GBP" => Currency::Gbp,
-            "KRW" => Currency::Krw,
-            "LUNA" => Currency::Luna,
-            "USD" => Currency::Usd,
-            "USDC" => Currency::Usdc,
-            "USDT" => Currency::Usdt,
-            other => Currency::Other(other.to_owned()),
-        })
-    }
-}
-
-impl<'de> Deserialize<'de> for Currency {
-    fn deserialize<D: de::Deserializer<'de>>(deserializer: D) -> Result<Self, D::Error> {
-        String::deserialize(deserializer)?
-            .parse()
-            .map_err(D::Error::custom)
-    }
-}
-
-impl Serialize for Currency {
-    fn serialize<S: ser::Serializer>(&self, serializer: S) -> Result<S::Ok, S::Error> {
-        self.to_string().serialize(serializer)
-    }
-}
-
-/// Trading pairs
-pub struct Pair(pub Currency, pub Currency);
-
-impl Pair {
-    /// Percent encode this pair (for inclusion in a URL)
-    pub fn percent_encode(&self) -> String {
-        utf8_percent_encode(&self.to_string(), NON_ALPHANUMERIC).to_string()
-    }
-}
-
-impl Display for Pair {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        write!(f, "{}/{}", self.0, self.1)
-    }
-}
-
-impl FromStr for Pair {
-    type Err = Error;
-
-    fn from_str(s: &str) -> Result<Self, Error> {
-        let pair: Vec<_> = s.split('/').collect();
-
-        if pair.len() != 2 {
-            fail!(ErrorKind::Parse, "malformed trading pair: {}", s);
-        }
-
-        Ok(Pair(pair[0].parse()?, pair[1].parse()?))
-    }
-}
-
-impl<'de> Deserialize<'de> for Pair {
-    fn deserialize<D: de::Deserializer<'de>>(deserializer: D) -> Result<Self, D::Error> {
-        String::deserialize(deserializer)?
-            .parse()
-            .map_err(D::Error::custom)
-    }
-}
-
-impl Serialize for Pair {
-    fn serialize<S: ser::Serializer>(&self, serializer: S) -> Result<S::Ok, S::Error> {
-        self.to_string().serialize(serializer)
-    }
-}
-
-/// Prices of currencies (internally represented as a `Decimal`)
-#[derive(Clone, Debug, Eq, PartialEq, PartialOrd, Ord)]
-pub struct Price(pub Decimal);
-
-impl Price {
-    /// Create a new price from a `Decimal`
-    pub(crate) fn new(decimal: Decimal) -> Result<Self, Error> {
-        if decimal.to_f32().is_none() || decimal.to_f64().is_none() {
-            fail!(ErrorKind::Parse, "price cannot be represented as float");
-        }
-
-        Ok(Price(decimal))
-    }
-}
-
-impl Display for Price {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        self.0.fmt(f)
-    }
-}
-
-impl FromStr for Price {
-    type Err = Error;
-
-    fn from_str(s: &str) -> Result<Self, Error> {
-        Self::new(s.parse()?)
-    }
-}
-
-impl<'de> Deserialize<'de> for Price {
-    fn deserialize<D: de::Deserializer<'de>>(deserializer: D) -> Result<Self, D::Error> {
-        String::deserialize(deserializer)?
-            .parse()
-            .map_err(D::Error::custom)
-    }
-}
-
-impl Serialize for Price {
-    fn serialize<S: ser::Serializer>(&self, serializer: S) -> Result<S::Ok, S::Error> {
-        self.to_string().serialize(serializer)
-    }
+    Ok((lowest_ask + highest_bid) / 2)
 }
